requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
mysql-connector-python>=8.3.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.2
Jinja2>=3.1.0
click>=8.1.3 